backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastapi==0.115.0",
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "lxml==5.3.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "python-dotenv==1.0.1",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "uvicorn[standard]>=0.42.0",
    "yfinance>=1.2.0",
]
